Fix parsing non-standard IconID in kdb files
This avoids the Entry/IconID parsing error for a non-standard icon ID 70
when opening an old `kdb` database. The fix is to parse an icon ID
either as a standard icon or a non-standard one with the original ID so
that the application is able to save it as is without unintentionally
modifying it. The UI displays the default key icon for any non-standard
icon mimicking KeePassX's behavior.